Changes
Changed personalization properties to be public since VSCode may try to suggest the underscore versions which looks gross in a call for the user.

Changed behavior or mouse-over. It no longer changes the selected row. It will still highlight the text of the row it's on and select that program if you click, but now if you use something like the Alt-Tab example, it won't select the wrong program when it bugged out and select the row the mouse is over when opened.

Due to previous change in behavior, removed saving of mouse position to not update selected row since keyboard and mouse selection are now independent of one another.

Added new color option for the row the mouse is hovering over (defaults to the same as selected row highlight text color).