Fixed mouse accuracy issue.
When I first added the OnMouseLeave message, it had inconsistency firing. The issue turned out to be the -Caption option in the gui. But I guess in adding the TrackMouseEvent portion, that seemed to fix the original problem as well but cause the mouse highlight row/close-button to be inaccurate.